+++ reverse-list/CHANGELOG.md
@@ -1,7 +1,11 @@
 # Revision history for reverse-list
 
-## 0.1.1 -- ????-??-??
+## 0.2.0 -- ????-??-??
 
+* Add `init`, `last`.
+* Add `Data.List.Cons` module and `Tsil` synonym for "symmetry".
+* Rename `Data.List.RList` to `Data.List.Snoc`
+* `RList.{safeLast,safeInit}`
 * `RList.toSet`
 
 ## 0.1.0 -- 2021-12-27